use mixin for status group function
