+++ plugins/tracker-resources/src/utils.ts
@@ -20,13 +20,11 @@ import core, {
   ApplyOperations,
   AttachedData,
   AttachedDoc,
-  CategoryType,
   Class,
   Collection,
   Doc,
   DocumentQuery,
   DocumentUpdate,
-  getObjectValue,
   Ref,
   SortingOrder,
   Space,
